Fix redux state and connect with backend
