+++ Client/package.json
@@ -14,16 +14,14 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@reduxjs/toolkit": "^1.9.5",
-    "axios": "^1.3.6",
-<<<<<<< HEAD
-=======
+    "axios": "^1.4.0",
     "bootstrap": "^5.2.3",
     "formik": "^2.2.9",
->>>>>>> e3323d096af88b1faf1dec4fa865d4ef6b13a1d7
     "next": "^13.3.1",
-    "react": "^18.2.0",  
+    "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^6.10.0",
     "yup": "^1.1.1"
   },
