+++ Client/src/App.jsx
@@ -1,15 +1,12 @@
 import { Routes, Route } from 'react-router-dom';
 import Landing from './views/Landing/Landing'
-import ReservationPanel from './views/ReservationPanel/ReservationPanel';
-//import createVehicle from "./components/createVehicle/createVehicle.jsx"
-import CreateVehicle from './components/CraeteVehicle/createVehicle.jsx'
 
 import CreateVehicle from './components/CraeteVehicle/createVehicle.jsx'
 import Dash from './components/dashboard/dash';
 import HomeView from './views/home/homeView';
 import ReservationView from './views/ReservationPanel/ReservationView';
 import CarsView from './views/ListCars/carsView';
-
+import ReservationPanel from './views/ReservationPanel/ReservationPanel';
 //import Home from "./components/Home/Home.jsx"
 
 
@@ -22,11 +19,10 @@ const App = () => {
       <Routes>
           <Route path='/' element={ <Landing /> } />
           {/* <Route path='/home' element={ <Home /> } /> */}
-          {/* <Route path='/reservation-panel' element={ <PaneldeReservacion /> } /> */}
+          <Route path='/reservation-panel' element={ <ReservationPanel /> } />
           <Route path='/createvehicle' element={ <CreateVehicle /> } />
           <Route path='/home' element={ <HomeView /> } /> 
           <Route path='/reservation-panel' element={ <ReservationView /> } /> 
-          <Route path='/createvehicle' element={ <CreateVehicle /> } />
            <Route path='/vehicles' element={<CarsView/>} /> 
           {/* <Route path='/about' element={ <About /> } /> */}
           {/* <Route path='/navbar' element={ <BarraNav /> } /> */}
