+++ Client/src/views/ReservationPanel/ReservationPanel.jsx
@@ -3,10 +3,6 @@ import { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import ParkingSpaceCard from '../../components/ParkingSpaceCard/ParkingSpaceCard';
 import { setCurrentPage } from '../../redux/features/parkingSpacesPagination/parkingSpacesPaginationSlice';
-import { getParkingLotById,
-    getLevelsByParkingLotId,
-    getParkingSpacesByParkingLotId
-} from '../../redux/features/parkingSpaces/parkingSpacesSlice';
 
 const ReservationPanel = () => {
     const dispatch = useDispatch();
@@ -21,9 +17,9 @@ const ReservationPanel = () => {
     const parkingSpacesInThisLevel = parking_spaces.filter(pS => pS.floorId === selectedLevel.id);
     const [filteredParkingSpaces, setfilteredParkingSpaces] = useState(parkingSpacesInThisLevel);
 
-    const availableParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'available');
-    const occupiedParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'occupied');
-    const reservedParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'reserved');
+    const availableParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'Disponible');
+    const occupiedParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'Ocupado');
+    const reservedParkingSpaces = parkingSpacesInThisLevel.filter(pS => pS.zone_status === 'Reservado');
 
     const currentPage = useSelector(state => state.parkingSpacesPagination.currentPage);
     const itemsPerPage = useSelector(state => state.parkingSpacesPagination.itemsPerPage);
@@ -32,11 +28,8 @@ const ReservationPanel = () => {
     const currentFilteredParkingSpaces = filteredParkingSpaces.slice(startIndex, endIndex);
 
     let selectedParkingSpace = useSelector(state => state.parkingSpaces.selectedParkingSpace);
-    useEffect(() => {
-        dispatch(getParkingLotById());
-        dispatch(getLevelsByParkingLotId());
-        dispatch(getParkingSpacesByParkingLotId());
 
+    useEffect(() => {
         if (Object.keys(selectedParkingSpace) !== 0) {
             setIsParkingSpaceSelected(true);
             console.log(isParkingSpaceSelected);
@@ -44,7 +37,7 @@ const ReservationPanel = () => {
         } else {
             setIsParkingSpaceSelected(false);
         }
-    }, [dispatch, selectedParkingSpace]);
+    }, [selectedParkingSpace]);
 
     const handleLevelSelection = (e) => {
         const selectedLevel = levels.find(level => level.name === e.target.textContent);
@@ -105,17 +98,17 @@ const ReservationPanel = () => {
                         <label htmlFor='parkingSpaceStatusFilter'>Estatus de la zona:</label>
                         <select id='parkingSpaceStatusFilter' onChange={ filterParkingSpaceByStatus } defaultValue=''>
                             <option value='' disabled>Estatus</option>
-                            <option value='available'>Disponibles</option>
-                            <option value='occupied'>Ocupados</option>
-                            <option value='reserved'>Reservados</option>
+                            <option value='Disponible'>Disponibles</option>
+                            <option value='Ocupado'>Ocupados</option>
+                            <option value='Reservado'>Reservados</option>
                         </select>
                     </div>
                     <div className={ styles.reservationPanel__parkingSpaces_filter }>
                         <label htmlFor='vehicleTypeFilter'>Zona para estacionar:</label>
                         <select id='vehicleTypeFilter' onChange={ filterByVehicleType } defaultValue=''>
                             <option value='' disabled>Tipo de vehículo</option>
-                            <option value='car'>carro</option>
-                            <option value='motorcycle'>motocicleta</option>
+                            <option value='Automovil'>carro</option>
+                            <option value='Motocicleta'>motocicleta</option>
                         </select>
                     </div>
                     <button type='button' onClick={ removeFilters }>Eliminar filtros</button>
@@ -141,8 +134,7 @@ const ReservationPanel = () => {
                             parking_space_id={ p_s.id }
 				    	    parking_space_label={ p_s.zone_number }
 				    	    vehicle_type={ p_s.vehicle_type }
-				    	    parking_space_status={ p_s.zone_status.charAt(0).toUpperCase() + 
-				    	        p_s.zone_status.slice(1) }
+				    	    parking_space_status={ p_s.zone_status }
 				    	/>
 				    )}
 				</div>
